Expose (Q)SPI pins to access flash memory
This patch exposes pins connected to the flash memory on
board (IS25LP064A) through the Board abstraction.

Although this does not introduce a high level library to access the
memory, it does allow for experimentation with standard SPI and QuadSPI
interfaces available in stm32h7xx-hal.

I managed to access the on board memory, read and write to it through
the HAL, example [1].

[1] https://github.com/richardeoin/stm32h7xx-hal/pull/3

Signed-off-by: Petr Horacek <hrck@protonmail.com>